fntest/fw: moved factories to new module
